requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
idna==3.3
prompt-toolkit==1.0.14
Pygments==2.10.0
PyInquirer==1.0.3
regex==2021.11.10
requests==2.26.0
six==1.16.0
SRTrain==2.0.2
urllib3==1.26.7
wcwidth==0.2.5
